Allow areas to be optionally open to all other users
In other words, you can retain ownership of an area but disable and
re-enable the protection aspect at will, via the area_open chat command.

By default, nothing is any different following this commit - all
existing areas are protected, as are new ones. But you can do (for
example) "area_open 1 yes" and if you're the owner of that area, it's
now possible for other users to interact there. "area_open 1 no" to
switch the protection back on.

Where there are sub-areas, the main owner and ALL sub-area owners at a
particulare location must have set the areas to open to disable
protection.